fixed the previous commit, start impliment operations and complete operation 8

+++ collegeManager/Student.cpp
@@ -1,6 +1,9 @@
 #include "Student.h"
 
-Student::Student(std::string _firstName, std::string _lastname, long _id, std::string _address, int _collegeStartYear, Department* _department) : 
+Student::Student(std::string _firstName, std::string _lastname, int _id, std::string _address, int _collegeStartYear, Department* _department) : 
 firstName(_firstName), lastName(_lastname), id(_id), address(_address), collegeStartYear(_collegeStartYear), departmentPtr(_department) {
-
 }
+
+//std::ostream& operator<<(std::ostream& os, const Student& student){
+//	os << "student first name: " << student.firstName
+//}
